Move sudoku solver to a new package
