+++ exercise_012_clustered_sudoku_solver/src/main/scala/akkapi/cluster/sudoku/ReductionRules.scala
@@ -1,4 +1,4 @@
-package akkapi.cluster
+package akkapi.cluster.sudoku
 
 object ReductionRules {
