+++ exercise_012_clustered_sudoku_solver/src/main/scala/akkapi/cluster/sudoku/SudokuDetailType.scala
@@ -1,4 +1,4 @@
-package akkapi.cluster
+package akkapi.cluster.sudoku
 
 sealed trait SudokoDetailType
 case class Row(id: Int) extends SudokoDetailType
