+++ exercise_012_clustered_sudoku_solver/src/test/scala/akkapi/cluster/sudoku/BaseAkkaSpec.scala
@@ -1,4 +1,4 @@
-package akkapi.cluster
+package akkapi.cluster.sudoku
 
 import akka.actor.testkit.typed.scaladsl.ActorTestKit
 import org.scalatest.BeforeAndAfterAll
