+++ exercise_012_clustered_sudoku_solver/src/test/scala/akkapi/cluster/sudoku/ReductionRuleSpec.scala
@@ -1,4 +1,4 @@
-package akkapi.cluster
+package akkapi.cluster.sudoku
 
 import org.scalatest.{Matchers, WordSpec}
